pyproject.toml
[tool.poetry]
name = "django-vuelidate"
version = "0.1.0"
description = "Generate vuelidate validation rules from Django models, forms, and serializers."
authors = ["Cody Kinsey"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.4"

[tool.poetry.dev-dependencies]
django-rest-framework = "^0.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
